rename SuggestedBlockDefinition to -Properties
