+++ packages/schema-blocks/src/core/validation/index.ts
@@ -3,6 +3,6 @@ import { BlockValidationResult } from "./BlockValidationResult";
 import { RequiredBlockOption } from "./RequiredBlockOption";
 import { RequiredBlock } from "./RequiredBlock";
 import { RecommendedBlock } from "./RecommendedBlock";
-import { SuggestedBlockDefinition } from "./SuggestedBlockDefinition";
+import { SuggestedBlockProperties } from "./SuggestedBlockProperties";
 
-export { BlockValidation, BlockValidationResult, RequiredBlockOption, RequiredBlock, RecommendedBlock, SuggestedBlockDefinition };
+export { BlockValidation, BlockValidationResult, RequiredBlockOption, RequiredBlock, RecommendedBlock, SuggestedBlockProperties };
